refactor(beacon): make contribution take an Rng
This allows us to remove the specific beacon implementation and specify the source of randomness via the CLI
