+++ powersoftau/src/lib.rs
@@ -1,5 +1,5 @@
 pub mod batched_accumulator;
+pub mod cli_common;
 pub mod keypair;
 pub mod parameters;
 pub mod utils;
-pub mod cli_common;
