+++ powersoftau/src/utils.rs
@@ -4,14 +4,85 @@ use blake2::{Blake2b, Digest};
 use byteorder::{BigEndian, ReadBytesExt};
 use generic_array::GenericArray;
 use rand::chacha::ChaChaRng;
-use rand::{Rand, Rng, SeedableRng};
+use rand::{OsRng, Rand, Rng, SeedableRng};
 
 use memmap::Mmap;
 use std::io::{self, Write};
 use std::sync::Arc;
 use typenum::consts::U64;
 
+use crypto::digest::Digest as CryptoDigest;
+use crypto::sha2::Sha256;
+
 use super::parameters::UseCompression;
+
+// Create an RNG based on a mixture of system randomness and user provided randomness
+pub fn user_system_randomness() -> Vec<u8> {
+    let mut system_rng = OsRng::new().unwrap();
+    let mut h = Blake2b::default();
+
+    // Gather 1024 bytes of entropy from the system
+    for _ in 0..1024 {
+        let r: u8 = system_rng.gen();
+        h.input(&[r]);
+    }
+
+    // Ask the user to provide some information for additional entropy
+    let mut user_input = String::new();
+    println!("Type some random text and press [ENTER] to provide additional entropy...");
+    std::io::stdin()
+        .read_line(&mut user_input)
+        .expect("expected to read some random text from the user");
+
+    // Hash it all up to make a seed
+    h.input(&user_input.as_bytes());
+    let arr: GenericArray<u8, U64> = h.result();
+    arr.to_vec()
+}
+
+#[allow(clippy::modulo_one)]
+pub fn beacon_randomness(mut beacon_hash: [u8; 32]) -> [u8; 32] {
+    // Performs 2^n hash iterations over it
+    const N: u64 = 10;
+
+    for i in 0..(1u64 << N) {
+        // Print 1024 of the interstitial states
+        // so that verification can be
+        // parallelized
+
+        if i % (1u64 << (N - 10)) == 0 {
+            print!("{}: ", i);
+            for b in beacon_hash.iter() {
+                print!("{:02x}", b);
+            }
+            println!();
+        }
+
+        let mut h = Sha256::new();
+        h.input(&beacon_hash);
+        h.result(&mut beacon_hash);
+    }
+
+    print!("Final result of beacon: ");
+    for b in beacon_hash.iter() {
+        print!("{:02x}", b);
+    }
+    println!();
+
+    beacon_hash
+}
+
+/// Interpret the first 32 bytes of the digest as 8 32-bit words
+pub fn get_rng(mut digest: &[u8]) -> impl Rng {
+    let mut seed = [0u32; 8];
+    for s in &mut seed {
+        *s = digest
+            .read_u32::<BigEndian>()
+            .expect("digest is large enough for this to work");
+    }
+
+    ChaChaRng::from_seed(&seed)
+}
 
 pub const fn num_bits<T>() -> usize {
     std::mem::size_of::<T>() * 8
